# Conflicts:
#	CHANGELOG.md
#	Cargo.toml
#	benches/README
#	benches/benchmarks.rs
#	deps_map.json
#	src/error.rs
#	src/executor/blockchain.rs
#	src/executor/crypto.rs
#	src/executor/deserialization.rs
#	src/executor/dump.rs
#	src/executor/engine/core.rs
#	src/executor/engine/data.rs
#	src/executor/exceptions.rs
#	src/executor/mod.rs
#	src/executor/rand.rs
#	src/lib.rs
#	src/smart_contract_info.rs
#	src/stack/continuation.rs
#	src/stack/integer/mod.rs
#	src/stack/integer/serialization/signed_big_endian.rs
#	src/stack/integer/serialization/signed_little_endian.rs
#	src/stack/integer/serialization/unsigned_big_endian.rs
#	src/stack/integer/serialization/unsigned_little_endian.rs
#	src/stack/mod.rs

+++ src/stack/integer/mod.rs
@@ -366,7 +366,7 @@ pub mod utils {
         if modpow2.is_zero() {
             return res
         }
-        (res + 1) as usize
+        res + 1
     }
 
     /// Perform in-place two's complement of the given digit iterator
